+++ scripts/generateDatLocaleData.js
@@ -1,11 +1,24 @@
 const fs = require('fs');
 const path = require('path');
+const { Grammar, Parser } = require('nearley');
+
+const inflection_grammar = Grammar.fromCompiled(
+  require('../src/grammars/generated/inflection.js')
+);
 
 const export_dir = path.join(__dirname, '../tmp/exports');
 const locale_files = fs.readdirSync(export_dir);
 
 const isExport = file => file.endsWith('.json');
 const underscore = s => s.toLowerCase();
+
+const fields = {
+  // [FieldName]: fieldValue => output
+  Name: name => parseInflection(name, inflection_grammar),
+  // skip emtpy strings
+  Inflection: id => (id === '' ? undefined : id)
+};
+const fields_with_inflection_rules = ['Name'];
 
 locale_files.filter(isExport).forEach(locale_file => {
   const locale = path.basename(locale_file, '.json');
@@ -18,7 +31,9 @@ locale_files.filter(isExport).forEach(locale_file => {
     const dat_name = path.basename(exported.filename, '.dat');
     const { header, data } = exported;
 
-    const output_header = ['Name'].reduce((partial_header, col_name) => {
+    const output_header = Object.keys(
+      fields
+    ).reduce((partial_header, col_name) => {
       partial_header[underscore(col_name)] = header.find(
         col => col.name === col_name
       );
@@ -31,7 +46,7 @@ locale_files.filter(isExport).forEach(locale_file => {
         output_header
       ).reduce((with_row, [name, info]) => {
         if (info !== undefined) {
-          with_row[name] = row[info.rowid];
+          with_row[name] = fields[info.name](row[info.rowid]);
         }
 
         return with_row;
@@ -46,3 +61,23 @@ locale_files.filter(isExport).forEach(locale_file => {
     );
   });
 });
+
+function parseInflection(raw, grammar) {
+  if (typeof raw !== 'string') return raw;
+
+  const parser = new Parser(grammar);
+
+  try {
+    parser.feed(raw);
+    const [[matches]] = parser.results;
+
+    if (matches === null) throw new Error('no result');
+
+    // convert to icu message syntax
+    return `{inflection, select, ${matches
+      .map(({ match, output }) => `${match} {${output}}`)
+      .join(' ')}}`;
+  } catch (err) {
+    return raw;
+  }
+}