+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased](https://github.com/eps1lon/poe-mods/compare/v0.5.0...HEAD)
+### Added
+- Gender and pluralization inflection rules for certain languages for Mods by using 
+  [ICU message syntax](http://userguide.icu-project.org/formatparse/messages).
 ### Fixed
 - wrong locale codes for simplified and traditional chinese `tw` and `cn` never
   were correct locale codes but country codes. The correct locales are `zh-tw`