+++ CHANGELOG.md
@@ -4,7 +4,9 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/eps1lon/poe-mods/compare/v0.5.0...HEAD)
+## [Unreleased](https://github.com/eps1lon/poe-mods/compare/v0.6.0...HEAD)
+
+## [0.6.0](https://github.com/eps1lon/poe-mods/compare/v0.5.0...v0.6.0) (2017-10-18)
 ### Added
 - Gender and pluralization inflection rules for certain languages for Mods by using 
   [ICU message syntax](http://userguide.icu-project.org/formatparse/messages).